Also test useFacetEffect
